pyproject.toml
[tool.poetry]
name = "emf-2024-eye"
version = "0.1.0"
description = ""
authors = ["Warren Moore <warren@wamo.uk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pygame = "^2.5.2"
pyopengl = "^3.1.7"
pyopengl-accelerate = "^3.1.7"
numpy = "^1.26.4"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
flake8-commas = "^4.0.0"
black = "^24.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
